moved ilg conversion methods to own module
