+++ test/test_parser.py
@@ -228,49 +228,6 @@ class TestParserFunctional(TestFSTOutput):
         self.assertEqual(result, expected)
     
 
-class TestStoryGlosser(unittest.TestCase):
-
-    def test_justLemma(self):
-        fst_ver, expected = ('w$an+VI', '___')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_Inflected(self):
-        fst_ver, expected = ('w$an+VI-3.II', '___-3.II')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_functionalInDict(self):
-        fst_ver, expected = ('n$ee+AUX', 'NEG')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_functionalInDictComplex(self):
-        fst_ver, expected = ('n$ee+AUX=EPIS=CN', 'NEG=EPIS=CN')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_looOblique(self):
-        fst_ver, expected = ('2PL+OBL', 'OBL-2PL.II')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_pronounIII(self):
-        fst_ver, expected = ('3.III+PRO', '3.III')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_demonstrative(self):
-        fst_ver, expected = ('PN=PROX+DEM', 'PN=DEM.PROX')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_demonstrativePl(self):
-        fst_ver, expected = ('ASSOC=PROX+DEM', 'ASSOC=DEM.PROX')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_quotative(self):
-        fst_ver, expected = ('1SG+QUOT', '1=QUOT')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-    def test_quotativeThirdPl(self):
-        fst_ver, expected = ('3PL+QUOT-3PL.INDP', '3=QUOT.3PL-3PL.INDP')
-        self.assertEqual(Parser.story_gloss(fst_ver), expected)
-
-
 # python -m unittest test
 # python -m unittest discover # all test files in current dir
 # python -m unittest discover -s tests # all test files in /tests
